Move the ExecAffectingOneRow to the postgres package
